build galimatias, htmlparser and langdetect through ant target
With this change, javac invocation made by ant should be equivalent to the way
it was built with build.py. Except that some non java files used to be in the
jar (because the script copied the content of src/ to classes/). However I
checked what the content of the built jar was when building with the original
build system of each module, and they didn't ship these files. So the content
of the jar produced with ant is, with regard to the non-java files, closer to
the one that is shipped in the jar.

The code in build.xml for ant is also now closer to the one of the pom.xml of
each module. Especially, in htmlparser, the pom.xml showed that there was an
additional step that wasn't managed by build.py (htmlparser-process-sources)
which is now done through ant.