remove useless <service> in war & bundle
They are auto-created properly if required. However, when building the jar including all dependencies
which content has been copied over into the final jar, it is required to override the providers
of the service org.relaxng.datatype.DatatypeLibraryFactory. In this case the providers are not
merged by the ant <jar> task and result in keeping the services/org.relaxng.datatype.DatatypeLibraryFactory
file only from validator.jar, and not considering the same file, but with different content
from jing.jar